+++ apps/dotcom/sync-worker/src/TLFileDurableObject.ts
